fix unresponsiveness of exchange tab while sending xml messages through plugins
